[smite] Remove transitive gevent imports from async worker
Extract gevent-free base modules so the async worker doesn't
transitively load gevent through shared osprey_worker code:

- grammar.py: threading.Semaphore instead of gevent.lock.Semaphore
- acking_contexts_base.py: BaseAckingContext, NoopAckingContext,
  VerdictsAckingContext without gevent (PubSub classes stay in
  acking_contexts.py)
- sources_provider_base.py: BaseSourcesProvider, StaticSourcesProvider
  without etcd/gevent
- external_service_utils_base.py: ExternalService, PlainExternalServiceAccessor
  without gevent.event.AsyncResult
- pigeon/client.py: local DOWN constant instead of importing service_watcher
- coordinator_input_stream.py: lazy Directory import (dead code in async path)
- execution_context.py: TYPE_CHECKING guard for ExternalServiceAccessor

All original modules re-export from base for backward compatibility.
Adds test_no_gevent_imports.py to enforce this going forward.
