+++ osprey_worker/src/osprey/engine/ast/grammar.py
@@ -6,7 +6,7 @@ from enum import Enum
 from pathlib import Path
 from typing import ClassVar, Dict, Optional, Sequence, TypeVar, Union
 
-from gevent.lock import Semaphore
+from threading import Semaphore
 
 # TODO: Uncomment logging when we have a logging system
 # from osprey.worker.ui_api.lib.osprey_shared.logging import get_logger
