+++ osprey_worker/src/osprey/worker/lib/sources_provider.py
@@ -1,39 +1,23 @@
-import abc
 import logging
-from typing import Callable, Dict, Optional
+from typing import Dict, Optional
 
 from osprey.engine.ast.sources import Sources
 from osprey.worker.lib.etcd import EtcdClient
 from osprey.worker.lib.etcd.dict import ReadOnlyEtcdDict
+from osprey.worker.lib.sources_provider_base import (
+    BaseSourcesProvider,
+    SourcesWatcherCallback,
+    StaticSourcesProvider,
+)
 from osprey.worker.lib.utils.input_stream_ready_signaler import InputStreamReadySignaler
 
-SourcesWatcherCallback = Callable[[], None]
-
-
-class BaseSourcesProvider(abc.ABC):
-    """Provides an interface to get the and be informed of current sources of rules which the rules engine should
-    evaluate"""
-
-    @abc.abstractmethod
-    def get_current_sources(self) -> Sources:
-        raise NotImplementedError
-
-    @abc.abstractmethod
-    def set_sources_watcher(self, callback: SourcesWatcherCallback) -> None:
-        raise NotImplementedError
-
-
-class StaticSourcesProvider(BaseSourcesProvider):
-    """Provides a static sources that won't change for the lifetime of the provider."""
-
-    def __init__(self, sources: Sources):
-        self._sources = sources
-
-    def get_current_sources(self) -> Sources:
-        return self._sources
-
-    def set_sources_watcher(self, callback: SourcesWatcherCallback) -> None:
-        return None
+# Re-export base classes for backward compatibility
+__all__ = [
+    'BaseSourcesProvider',
+    'StaticSourcesProvider',
+    'SourcesWatcherCallback',
+    'EtcdSourcesProvider',
+]
 
 
 class EtcdSourcesProvider(BaseSourcesProvider):
